Fixed sizeof expression in IOSLongArray. 	Change on 2013/01/31 by tball <tball@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=41899991

+++ jre_emul/Classes/IOSLongArray.m
@@ -40,7 +40,7 @@
 - (id)initWithLongs:(const long long *)longs count:(NSUInteger)count {
   if ((self = [self initWithLength:count])) {
     if (longs != nil) {
-      memcpy(buffer_, longs, count * sizeof(long));
+      memcpy(buffer_, longs, count * sizeof(long long));
     }
   }
   return self;
